requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pymysql
alembic
motor
cryptography
python-multipart
passlib
bcrypt
python-dotenv